sunil-console/ai-service/requirements.txt
# URackIT AI Service - Python Dependencies
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# OpenAI Agent SDK
openai-agents==0.0.15

# Database
psycopg2-binary==2.9.10
requests==2.32.3

# Knowledge Base / Vector Store
chromadb==0.5.23

# WebSocket Support
websockets==14.1

# Twilio SIP Integration
twilio==9.4.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
